Add a memory leak testing script
I've been manually running this as:

    RUBY_HEAP_SLOTS_GROWTH_FACTOR=1 RUBY_HEAP_MIN_SLOTS=1 RUBY_HEAP_SLOTS_INCREMENT=10 RUBY_HEAP_SLOTS_GROWTH_FACTOR=.1 ITERATIONS=75000 rspec -I lib -I profile profile/memory_leaks.rb

The ruby GC settings are that way to try and make ruby increase memory really
slowly, so that we can get more accurate info on what's going on.

ITERATIONS can be whatever, but I've been running it with a really large number
so that I can be more comfortable with the results.