Reintroduce mirror memoization.
The problem with memoization was that log_inspect triggered it right
after executing the complete handler. Typhoeus resets the easy in order
to reuse it. So after resetting log_inspect triggered the memoization
although it only was an empty easy. We are now logging before the
complete handler which is safer.