Do not depend on mime-types gem any more.
In order to not block users from updating to a recent version of
`mime-types` it won't be required for ethon any more. It'll be
recognized and used though if it is available. Putting `gem
'mime-types'` in your gemfile or having `mime-types` gem installed is
enough to make it work. Fixes #76.