Add support for :file and :writedata options to easy
These options take a file path as a string, and will write
the body of the response directly to the passed path, instead
of pulling it into the heap.

When passed, these options set :writefunction to NULL, so that
Curl knows to write directly to a file.

Easy manages the file handle, and closes it automatically in
on_complete.