Add support for sending empty header values.
If an empty string is passed as the value of a header, it now gets sent
as an empty HTTP header, rather than omitting the header. A nil value
will still omit the header.

This requires curl 7.23.0 or higher, where the ability to send empty
header values (via a semicolon terminator) was added. If using an older
version of curl without this support, then curl's previous behavior of
omitting empty headers remains the same.