Avoid mutating strings to support `--enable-frozen-string-literal`.
Prior to this commit, projects that use Ethon were not able to use this
option because Ethon mutated a string that started as a literal. This
commit fixes the issue by initializing the mutable string with
`String.new` instead of `''`.

In addition, this updates Ethon to enforce this by adding a
`# frozen_string_literal: true` comment to the top of all ruby files.

Ideally, we could have set `RUBYOPT=--enable-frozen-string-literal`
in the CI builds for the project, but when I tried that I learned
that mustermann (a dependency of Sinatra, which Ethon uses in its
test suite to boot a web server) mutates some string literals. So
the best we can do until mustermann is fixed is to enforce it via
comments at the top of each Ruby source file.

Luckily, rubocop provides a way to add it automatically, which I used
for this:

```
$ rubocop --only Style/FrozenStringLiteralComment -A
```