Use generic `Rackup::Handler.default`
Move from specific `Rack::Handler::WEBrick` to more generic
`Rackup::Handler.default`. This chooses the best available handler
automatically. If more handlers are available, `RACKUP_HANDLER`
environment variable could be used to override the default.

ATM, Gemfile still specifies WEBrick, so this is the one which will be
used. But replacing WEBrick by Puma should be as easy as updating the
`Gemfile`